[fix] don't delete loss.csv log file

+++ src/utils/utils.py
@@ -166,7 +166,7 @@ def clean_model_weights(best_epoch, fold, checkpoint_path, logger):
     # only keep files ending on correct fold
     files = [file for file in files if file.split(".")[0].endswith(str(fold)) and file.startswith("checkpoint")]
     for file in files:
-        if f"checkpoint_{best_epoch}_fold-{fold}" not in file:
+        if f"checkpoint_{best_epoch}_fold-{fold}" not in file and file.endswith(".tar"):
             os.remove(os.path.join(checkpoint_path, file))
     logger.info(f"Removed all checkpoint weights except the best one: checkpoint_{best_epoch}_fold-{fold}")
